+++ spec/spec_helper.rb
@@ -18,4 +18,6 @@ RSpec.configure do |config|
   config.order = :random
   config.filter_run focus: true
   config.run_all_when_everything_filtered = true
+
+  config.example_status_persistence_file_path = "tmp/.rspec-status"
 end